+++ weapon.java
@@ -8,12 +8,12 @@
 public class weapon extends Item
 {
     // instance variables - replace the example below with your own
-    private float power;
+    private double power;
 
     /**
      * Constructor for objects of class weapon
      */
-    public weapon(String name,float weight,float power,boolean canTake)
+    public weapon(String name,double weight,double power,boolean canTake)
     {
         super(name,weight,canTake);
         this.power = power;
@@ -25,7 +25,7 @@ public class weapon extends Item
      * @param  y   a sample parameter for a method
      * @return     the sum of x and y 
      */
-    public float getPower()
+    public double getPower()
     {
         return power;
     }
